Split admin playlist manager modules
