+++ frontend/components/admin/PlaylistsManager.tsx
@@ -11,378 +11,37 @@ import {
   useTransition,
   type DragEvent,
   type FormEvent,
-  type ReactNode,
 } from 'react';
 import clsx from 'clsx';
 import { Button } from '@/components/ui/Button';
 import { authFetch } from '@/lib/authFetch';
-import { getExampleFamilyDescriptor, getExampleFamilyIds, getExampleFamilyModelSlugs } from '@/lib/model-families';
-
-type PlaylistKind = 'core' | 'model' | 'draft';
-type PlaylistSurfaceRole = 'starter' | 'examplesHub' | 'family' | 'model' | 'draft' | 'other';
-type PlaylistSurfaceStatus = 'ready' | 'missing' | 'empty' | 'other';
-type PlaylistGroup = 'runtime' | 'family' | 'model' | 'draft';
-
-type PlaylistSummary = {
-  id: string;
-  slug: string;
-  name: string;
-  description: string | null;
-  isPublic: boolean;
-  createdAt: string;
-  updatedAt: string;
-  itemCount: number;
-  siteVisibleCount: number;
-  withVideoAssetCount: number;
-  lastAddedAt: string | null;
-  kind: PlaylistKind;
-  isLocked: boolean;
-  usageTargets: string[];
-  surfaceRole: PlaylistSurfaceRole;
-  surfaceStatus: PlaylistSurfaceStatus;
-  familyId: string | null;
-  modelSlug: string | null;
-  helperText: string | null;
-  drivesRoute: string | null;
-  fallbackModelSlugs: string[];
-};
-
-type PlaylistItemRecord = {
-  playlistId: string;
-  videoId: string;
-  orderIndex: number;
-  pinned: boolean;
-  createdAt: string;
-  thumbUrl?: string | null;
-  videoUrl?: string | null;
-  engineLabel?: string | null;
-  aspectRatio?: string | null;
-  prompt?: string | null;
-  durationSec?: number | null;
-  visibility: 'public' | 'private';
-  indexable: boolean;
-  isPublishedOnSite: boolean;
-};
-
-type PlaylistsManagerProps = {
-  initialPlaylists: PlaylistSummary[];
-  initialPlaylistId: string | null;
-  initialItems: PlaylistItemRecord[];
-  embedded?: boolean;
-  className?: string;
-};
-
-type EditablePlaylist = PlaylistSummary & { dirty?: boolean; loading?: boolean };
-type DropPlacement = 'before' | 'after';
-type FamilyPlaylistHelperCard = {
-  familyId: string;
-  label: string;
-  slug: string;
-  route: string;
-  fallbackModelSlugs: string[];
-  helperText: string;
-  status: 'ready' | 'missing' | 'empty';
-  playlistId: string | null;
-};
-
-const PLACEHOLDER_MAP: Record<string, string> = {
-  '9:16': '/assets/frames/thumb-9x16.svg',
-  '16:9': '/assets/frames/thumb-16x9.svg',
-  '1:1': '/assets/frames/thumb-1x1.svg',
-};
-
-const GROUP_LABELS: Record<PlaylistGroup, string> = {
-  runtime: 'Runtime surfaces',
-  family: 'Family playlists',
-  model: 'Model playlists',
-  draft: 'Draft / misc',
-};
-
-function getPlaceholderThumb(aspectRatio?: string | null): string {
-  const key = (aspectRatio ?? '').trim();
-  return PLACEHOLDER_MAP[key] ?? '/assets/frames/thumb-16x9.svg';
-}
-
-function formatDate(value?: string | null): string {
-  if (!value) return 'Never';
-  try {
-    return new Intl.DateTimeFormat('en-US', {
-      dateStyle: 'medium',
-      timeStyle: 'short',
-    }).format(new Date(value));
-  } catch {
-    return value;
-  }
-}
-
-function slugify(value: string): string {
-  return value
-    .trim()
-    .toLowerCase()
-    .replace(/[^a-z0-9]+/g, '-')
-    .replace(/^-+|-+$/g, '')
-    .slice(0, 80);
-}
-
-function compareDatesDescending(a?: string | null, b?: string | null) {
-  const aTime = a ? Date.parse(a) : Number.NaN;
-  const bTime = b ? Date.parse(b) : Number.NaN;
-  const safeA = Number.isFinite(aTime) ? aTime : 0;
-  const safeB = Number.isFinite(bTime) ? bTime : 0;
-  return safeB - safeA;
-}
-
-function getPlaylistPriority(playlist: Pick<EditablePlaylist, 'surfaceRole'>) {
-  if (playlist.surfaceRole === 'starter') return 0;
-  if (playlist.surfaceRole === 'examplesHub') return 1;
-  if (playlist.surfaceRole === 'family') return 2;
-  if (playlist.surfaceRole === 'model') return 3;
-  if (playlist.surfaceRole === 'other') return 4;
-  return 5;
-}
-
-function comparePlaylists(a: EditablePlaylist, b: EditablePlaylist) {
-  const priorityDelta = getPlaylistPriority(a) - getPlaylistPriority(b);
-  if (priorityDelta !== 0) return priorityDelta;
-  if (a.itemCount !== b.itemCount) return b.itemCount - a.itemCount;
-  const lastAddedDelta = compareDatesDescending(a.lastAddedAt, b.lastAddedAt);
-  if (lastAddedDelta !== 0) return lastAddedDelta;
-  return a.name.localeCompare(b.name) || a.slug.localeCompare(b.slug);
-}
-
-function sortPlaylists(playlists: EditablePlaylist[]): EditablePlaylist[] {
-  return [...playlists].sort(comparePlaylists);
-}
-
-function getUsageLabel(target: string): string {
-  if (target === 'starter-tab') return 'Starter tab';
-  if (target === 'examples-hub') return 'Examples hub';
-  if (target.startsWith('family-page:')) {
-    return `Family page · ${target.slice('family-page:'.length)}`;
-  }
-  if (target.startsWith('family-fallback:')) {
-    return `Family fallback · ${target.slice('family-fallback:'.length)}`;
-  }
-  if (target.startsWith('model-page:')) {
-    return `Model page · ${target.slice('model-page:'.length)}`;
-  }
-  return target;
-}
-
-function summarizePlaylist(playlist: PlaylistSummary): string {
-  if (!playlist.itemCount) return 'Empty collection';
-  return `${playlist.itemCount} items · ${playlist.siteVisibleCount} live · ${playlist.withVideoAssetCount} with video`;
-}
-
-function buildPlaylistUpdateFromItems(playlist: EditablePlaylist, items: PlaylistItemRecord[]): EditablePlaylist {
-  const lastAddedAt = items.reduce<string | null>((latest, item) => {
-    if (!latest) return item.createdAt;
-    return compareDatesDescending(item.createdAt, latest) < 0 ? item.createdAt : latest;
-  }, null);
-
-  return {
-    ...playlist,
-    itemCount: items.length,
-    siteVisibleCount: items.filter((item) => item.isPublishedOnSite).length,
-    withVideoAssetCount: items.filter((item) => Boolean(item.videoUrl)).length,
-    lastAddedAt,
-  };
-}
-
-function getPlaylistGroup(playlist: Pick<PlaylistSummary, 'surfaceRole'>): PlaylistGroup {
-  if (playlist.surfaceRole === 'starter' || playlist.surfaceRole === 'examplesHub') {
-    return 'runtime';
-  }
-  if (playlist.surfaceRole === 'family') {
-    return 'family';
-  }
-  if (playlist.surfaceRole === 'model') {
-    return 'model';
-  }
-  return 'draft';
-}
-
-function buildFamilyHelpers(playlists: PlaylistSummary[]): FamilyPlaylistHelperCard[] {
-  const byFamilyId = new Map<string, PlaylistSummary>();
-  playlists.forEach((playlist) => {
-    if (playlist.surfaceRole === 'family' && playlist.familyId) {
-      byFamilyId.set(playlist.familyId, playlist);
-    }
-  });
-
-  return getExampleFamilyIds().map((familyId) => {
-    const descriptor = getExampleFamilyDescriptor(familyId);
-    const playlist = byFamilyId.get(familyId) ?? null;
-    const fallbackModelSlugs = getExampleFamilyModelSlugs(familyId);
-    return {
-      familyId,
-      label: descriptor?.label ?? familyId,
-      slug: `family-${familyId}`,
-      route: `/examples/${familyId}`,
-      fallbackModelSlugs,
-      helperText: `Feeds /examples/${familyId}. Auto-fills from ${
-        fallbackModelSlugs.length
-          ? `${fallbackModelSlugs.map((modelSlug) => `examples-${modelSlug}`).join(', ')} then examples`
-          : 'examples'
-      }.`,
-      status: playlist ? (playlist.itemCount > 0 ? 'ready' : 'empty') : 'missing',
-      playlistId: playlist?.id ?? null,
-    };
-  });
-}
-
-function sortItemsForDisplay(items: PlaylistItemRecord[]): PlaylistItemRecord[] {
-  return [...items].sort((a, b) => {
-    if (a.orderIndex !== b.orderIndex) {
-      return b.orderIndex - a.orderIndex;
-    }
-    return compareDatesDescending(a.createdAt, b.createdAt);
-  });
-}
-
-function getSurfaceRoleLabel(role: PlaylistSurfaceRole): string {
-  switch (role) {
-    case 'starter':
-      return 'Starter surface';
-    case 'examplesHub':
-      return 'Examples hub';
-    case 'family':
-      return 'Family playlist';
-    case 'model':
-      return 'Model playlist';
-    case 'draft':
-      return 'Draft';
-    case 'other':
-    default:
-      return 'Misc collection';
-  }
-}
-
-function getSurfaceRoleTone(role: PlaylistSurfaceRole): 'neutral' | 'ok' | 'warn' {
-  if (role === 'starter' || role === 'examplesHub') return 'ok';
-  if (role === 'family' || role === 'model') return 'neutral';
-  return 'warn';
-}
-
-function getSurfaceStatusTone(status: PlaylistSurfaceStatus | FamilyPlaylistHelperCard['status']): 'neutral' | 'ok' | 'warn' {
-  if (status === 'ready') return 'ok';
-  if (status === 'missing' || status === 'empty') return 'warn';
-  return 'neutral';
-}
-
-function getSurfaceStatusLabel(status: PlaylistSurfaceStatus | FamilyPlaylistHelperCard['status']): string {
-  if (status === 'ready') return 'Ready';
-  if (status === 'missing') return 'Missing';
-  if (status === 'empty') return 'Empty';
-  return 'General';
-}
-
-function buildHelperFallbackLabel(modelSlug: string): string {
-  return `examples-${modelSlug}`;
-}
-
-function StatusPill({
-  tone,
-  children,
-}: {
-  tone: 'neutral' | 'ok' | 'warn';
-  children: ReactNode;
-}) {
-  const classes =
-    tone === 'ok'
-      ? 'border-emerald-500/30 bg-emerald-500/10 text-emerald-700'
-      : tone === 'warn'
-        ? 'border-warning-border/60 bg-warning-bg/20 text-warning'
-        : 'border-border bg-bg text-text-secondary';
-
-  return <span className={clsx('inline-flex rounded-full border px-2 py-1 text-[11px] font-semibold', classes)}>{children}</span>;
-}
-
-function PlaylistRailCard({
-  playlist,
-  isActive,
-  onSelect,
-}: {
-  playlist: EditablePlaylist;
-  isActive: boolean;
-  onSelect: (playlistId: string) => void;
-}) {
-  return (
-    <button
-      type="button"
-      onClick={() => onSelect(playlist.id)}
-      className={clsx(
-        'w-full rounded-card border p-4 text-left transition focus-visible:outline-none focus-visible:ring-2 focus-visible:ring-ring focus-visible:ring-offset-2 focus-visible:ring-offset-bg',
-        isActive
-          ? 'border-brand bg-surface-2 shadow-card'
-          : 'border-hairline bg-surface hover:border-text-muted hover:bg-surface-2'
-      )}
-    >
-      <div className="flex items-start justify-between gap-3">
-        <div className="min-w-0 space-y-1">
-          <p className="truncate text-sm font-semibold text-text-primary">{playlist.name}</p>
-          <p className="truncate text-xs text-text-muted">{playlist.slug}</p>
-        </div>
-        <div className="flex shrink-0 flex-wrap justify-end gap-2">
-          <StatusPill tone={getSurfaceRoleTone(playlist.surfaceRole)}>{getSurfaceRoleLabel(playlist.surfaceRole)}</StatusPill>
-          <StatusPill tone={getSurfaceStatusTone(playlist.surfaceStatus)}>{getSurfaceStatusLabel(playlist.surfaceStatus)}</StatusPill>
-        </div>
-      </div>
-      <p className="mt-3 text-xs text-text-secondary">{summarizePlaylist(playlist)}</p>
-      {playlist.helperText ? <p className="mt-2 text-[11px] leading-5 text-text-muted">{playlist.helperText}</p> : null}
-      <p className="mt-2 text-[11px] text-text-muted">Last added: {formatDate(playlist.lastAddedAt)}</p>
-      <div className="mt-3 flex flex-wrap gap-2">
-        {playlist.usageTargets.map((target) => (
-          <StatusPill key={target} tone={playlist.surfaceRole === 'starter' || playlist.surfaceRole === 'examplesHub' ? 'ok' : 'neutral'}>
-            {getUsageLabel(target)}
-          </StatusPill>
-        ))}
-        {playlist.isLocked ? <StatusPill tone="neutral">Locked</StatusPill> : null}
-      </div>
-    </button>
-  );
-}
-
-function MissingFamilyCard({
-  helper,
-  onCreate,
-  onSeed,
-  pending,
-}: {
-  helper: FamilyPlaylistHelperCard;
-  onCreate: (familyId: string) => void;
-  onSeed: (familyId: string) => void;
-  pending: boolean;
-}) {
-  return (
-    <div className="rounded-card border border-dashed border-hairline bg-surface p-4">
-      <div className="flex items-start justify-between gap-3">
-        <div className="min-w-0 space-y-1">
-          <p className="truncate text-sm font-semibold text-text-primary">{helper.label}</p>
-          <p className="truncate text-xs text-text-muted">{helper.slug}</p>
-        </div>
-        <StatusPill tone={getSurfaceStatusTone(helper.status)}>{getSurfaceStatusLabel(helper.status)}</StatusPill>
-      </div>
-      <p className="mt-3 text-xs text-text-secondary">{helper.helperText}</p>
-      <div className="mt-3 flex flex-wrap gap-2">
-        <StatusPill tone="neutral">{helper.route}</StatusPill>
-        {helper.fallbackModelSlugs.map((modelSlug) => (
-          <StatusPill key={modelSlug} tone="neutral">
-            {buildHelperFallbackLabel(modelSlug)}
-          </StatusPill>
-        ))}
-      </div>
-      <div className="mt-4 flex flex-wrap gap-2">
-        <Button type="button" size="sm" variant="outline" onClick={() => onCreate(helper.familyId)} disabled={pending}>
-          Create empty
-        </Button>
-        <Button type="button" size="sm" onClick={() => onSeed(helper.familyId)} disabled={pending}>
-          Seed from auto
-        </Button>
-      </div>
-    </div>
-  );
-}
+import { MissingFamilyCard } from '@/components/admin/playlists/MissingFamilyCard';
+import { PlaylistRailCard } from '@/components/admin/playlists/PlaylistRailCard';
+import { StatusPill } from '@/components/admin/playlists/PlaylistStatusPill';
+import {
+  GROUP_LABELS,
+  buildFamilyHelpers,
+  buildHelperFallbackLabel,
+  buildPlaylistUpdateFromItems,
+  formatDate,
+  getPlaceholderThumb,
+  getPlaylistGroup,
+  getSurfaceRoleLabel,
+  getSurfaceRoleTone,
+  getSurfaceStatusLabel,
+  getSurfaceStatusTone,
+  getUsageLabel,
+  slugify,
+  sortItemsForDisplay,
+  sortPlaylists,
+} from '@/components/admin/playlists/playlist-helpers';
+import type {
+  DropPlacement,
+  EditablePlaylist,
+  PlaylistItemRecord,
+  PlaylistsManagerProps,
+  PlaylistSummary,
+} from '@/components/admin/playlists/playlist-types';
 
 export function PlaylistsManager({
   initialPlaylists,
